Move the ParameterProvider interface to the parametric.providers package
