+++ BCM-Universal/src/main/java/me/fixeddev/bcm/parametric/providers/ParameterProviderRegistryImpl.java
@@ -1,6 +1,5 @@
 package me.fixeddev.bcm.parametric.providers;
 
-import me.fixeddev.bcm.parametric.ParameterProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
