+++ BCM-Universal/src/main/java/me/fixeddev/bcm/parametric/providers/defaults/NamespaceProvider.java
@@ -3,11 +3,10 @@ package me.fixeddev.bcm.parametric.providers.defaults;
 import me.fixeddev.bcm.basic.ArgumentStack;
 import me.fixeddev.bcm.basic.Namespace;
 import me.fixeddev.bcm.basic.exceptions.NoMoreArgumentsException;
-import me.fixeddev.bcm.parametric.ParameterProvider;
+import me.fixeddev.bcm.parametric.providers.ParameterProvider;
 import me.fixeddev.bcm.basic.exceptions.ArgumentsParseException;
 
 import java.lang.annotation.Annotation;
-import java.util.List;
 
 public class NamespaceProvider implements ParameterProvider<Namespace> {
     @Override
